fixed issue with app not launching

+++ app/src/main/java/com/simple/cats/presentation/views/pages/BreedFactsPage.kt
@@ -166,7 +166,7 @@ fun SearchBarBreedFacts(
             onSearch = {
                 searchQuery?.let { query ->
 
-                    breedFactsViewModel.getBreed(query)
+                    breedFactsViewModel.getBreed(query.toInt())
                     onSearch(query)
                 }
                 focusManager.clearFocus()
